Separated Labels and their wdgets into files
