+++ shortcut_composer/templates/pie_menu_utils/label_widgets/__init__.py
@@ -1,0 +1,4 @@
+from .create_label_widget import create_label_widget
+from .label_widget import LabelWidget
+
+__all__ = ["create_label_widget", "LabelWidget"]
